feat(settings): add sidebar nav for settings + agent job description
Splits the long settings page into a left-rail navigation with one
section visible at a time. Each settings area lives in its own file
under app/src/components/settings/sections/.

Pulls Connect phone (with QR) out of the global sidebar and into the
settings rail behind a BETA pill, so the QR, the connection state, and
the disconnect-all-phones reset live together.

Reuses the same SidebarSectionNav for the agent Job Description tab,
replacing the pill row with Instructions / Actions (LibraryBig) /
Learnings (Brain).
