+++ app/src/components/tabs/job-description-tab.tsx
@@ -1,4 +1,6 @@
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
+import { useTranslation } from "react-i18next";
+import { FileText, LibraryBig, Brain } from "lucide-react";
 import { SkillDetailPage } from "@houston-ai/skills";
 import {
   useInstructions,
@@ -11,24 +13,29 @@ import {
 import type { TabProps } from "../../lib/types";
 import { useUIStore } from "../../stores/ui";
 import { LearningsContent } from "./learnings-content";
-import {
-  InstructionsContent,
-  SubTabPills,
-  type SubTab,
-} from "./job-description-parts";
+import { InstructionsContent, type SubTab } from "./job-description-parts";
 import { ActionsContent } from "./actions-content";
 import { useActionSurface } from "./use-action-surface";
+import {
+  SidebarSectionNav,
+  type SidebarSectionItem,
+} from "../shared/sidebar-section-nav";
 
 export default function JobDescriptionTab({ agent }: TabProps) {
+  const { t } = useTranslation("agents");
   const path = agent.folderPath;
   const actions = useActionSurface(path);
   const [activeTab, setActiveTab] = useState<SubTab>("instructions");
   const targetTab = useUIStore((s) => s.jobDescriptionTarget);
   const setTargetTab = useUIStore((s) => s.setJobDescriptionTarget);
 
-  // Instructions (CLAUDE.md)
   const { data: instructions } = useInstructions(path);
   const saveInstructions = useSaveInstructions(path);
+
+  const { data: learningsData } = useLearnings(path);
+  const addLearning = useAddLearning(path);
+  const removeLearning = useRemoveLearning(path);
+  const updateLearning = useUpdateLearning(path);
 
   useEffect(() => {
     if (!targetTab) return;
@@ -37,13 +44,16 @@ export default function JobDescriptionTab({ agent }: TabProps) {
     setTargetTab(null);
   }, [targetTab, setTargetTab, actions.clearSelectedSkill]);
 
-  // Learnings
-  const { data: learningsData } = useLearnings(path);
-  const addLearning = useAddLearning(path);
-  const removeLearning = useRemoveLearning(path);
-  const updateLearning = useUpdateLearning(path);
+  const items = useMemo<SidebarSectionItem<SubTab>[]>(
+    () => [
+      { id: "instructions", label: t("subTabs.instructions"), icon: FileText },
+      { id: "skills", label: t("subTabs.skills"), icon: LibraryBig },
+      { id: "learnings", label: t("subTabs.learnings"), icon: Brain },
+    ],
+    [t],
+  );
 
-  // Actions sub-tab in detail mode takes over the whole pane (its own header).
+  // Skill detail view takes over the whole pane.
   if (activeTab === "skills" && actions.selectedSkill) {
     return (
       <SkillDetailPage
@@ -57,10 +67,13 @@ export default function JobDescriptionTab({ agent }: TabProps) {
   }
 
   return (
-    <div className="flex-1 flex flex-col min-h-0 bg-background">
-      <SubTabPills activeTab={activeTab} onChange={setActiveTab} />
-
-      {/* Body: each sub-tab decides its own layout. */}
+    <div className="flex-1 flex min-h-0 bg-background">
+      <SidebarSectionNav
+        ariaLabel={agent.name}
+        items={items}
+        active={activeTab}
+        onSelect={setActiveTab}
+      />
       <div className="flex-1 min-h-0 overflow-y-auto flex flex-col">
         {activeTab === "instructions" && (
           <InstructionsContent
@@ -72,7 +85,7 @@ export default function JobDescriptionTab({ agent }: TabProps) {
         )}
 
         {activeTab === "skills" && (
-          <div className="max-w-3xl mx-auto w-full px-6 pb-12 pt-2 flex-1 flex flex-col">
+          <div className="max-w-3xl mx-auto w-full px-6 pb-12 pt-6 flex-1 flex flex-col">
             <ActionsContent
               skills={actions.skills}
               loading={actions.skillsLoading}
