+++ app/src/locales/en/shell.json
@@ -5,7 +5,6 @@
     "settings": "Settings",
     "yourAgents": "Your Agents",
     "selectWorkspace": "Select workspace",
-    "connectPhone": "Connect phone",
     "addAgent": "New agent",
     "agentActions": "Agent actions",
     "color": "Color",
@@ -105,17 +104,5 @@
   },
   "engineGate": {
     "starting": "Starting Houston engine…"
-  },
-  "pairDevice": {
-    "title": "Connect your phone",
-    "description": "Scan this code to manage Houston from your iPhone.",
-    "loading": "Getting ready to connect your phone…",
-    "connectingStatus": "Connecting to the phone service. This usually takes a few seconds.",
-    "keepMacAwake": "Keep this Mac awake while you use Houston on your phone.",
-    "alwaysOnHint": "For 24/7 access, <emph>Always On</emph> runs Houston on our servers.",
-    "errors": {
-      "noInternet": "Houston needs an internet connection the first time you open the app. Please connect and reopen Houston to enable phone pairing.",
-      "generic": "Couldn't prepare a pairing code. Please try again."
-    }
   }
 }
