+++ app/src/locales/es/shell.json
@@ -5,7 +5,6 @@
     "settings": "Configuración",
     "yourAgents": "Tus agentes",
     "selectWorkspace": "Elige un espacio de trabajo",
-    "connectPhone": "Conectar celular",
     "addAgent": "Nuevo agente",
     "agentActions": "Acciones del agente",
     "color": "Color",
@@ -105,17 +104,5 @@
   },
   "engineGate": {
     "starting": "Iniciando el motor de Houston…"
-  },
-  "pairDevice": {
-    "title": "Conecta tu celular",
-    "description": "Escanea este código para manejar Houston desde tu iPhone.",
-    "loading": "Preparando la conexión con tu celular…",
-    "connectingStatus": "Conectando con el servicio del celular. Suele tomar unos segundos.",
-    "keepMacAwake": "Mantén este Mac despierto mientras usas Houston en tu celular.",
-    "alwaysOnHint": "Para acceso 24/7, <emph>Always On</emph> ejecuta Houston en nuestros servidores.",
-    "errors": {
-      "noInternet": "Houston necesita conexión a internet la primera vez que lo abres. Conéctate y reabre Houston para habilitar el emparejamiento del celular.",
-      "generic": "No se pudo generar un código de emparejamiento. Inténtalo de nuevo."
-    }
   }
 }
